requirements.txt
bash
loguru
PyYAML
fastcore
fastapi
pandas
gspread
gspread_dataframe
oauth2client
google-api-python-client